+++ src/config/config.c
@@ -20,7 +20,7 @@ ShellConfig shell_config = {
     .prompt = "Shell > ",
     .aliases = {},
     .alias_count = 0,
-    .timeout = 10000,
+    .timeout = -1,
     .show_full_cmd_path = false,
     .show_full_alias_cmd = false,
     .show_exit_code = false,
@@ -38,7 +38,37 @@ ShellConfig shell_config = {
     },
     .env_count = 0, // gets set at runtime in init shell config
     .allow_env_override = false,
+    .terminal_mode = TERM_MODE_NORMAL,
+    .enable_history_file = true,
+    .enable_history = true,
+    .max_history_len = 4096,
+    .history_file = NULL,
 };
+
+static bool parse_string_to_int(const char *value, int *out) {
+    if (!value) return false;
+
+    char *endptr;
+    errno = 0;
+    long val = strtol(value, &endptr, 10);
+
+    if (errno == ERANGE || endptr == value || *endptr != '\0') {
+        return false;
+    }
+
+    if (val < INT_MIN || val > INT_MAX) {
+        return false;
+    }
+
+    *out = (int)val;
+    return true;
+}
+
+static const char *int_to_string(int num) {
+    static char buf[32];
+    snprintf(buf, sizeof(buf), "%d", num);
+    return buf;
+}
 
 bool set_info_color(const char *value) {
     return set_color("info", value);
@@ -104,9 +134,18 @@ const char *get_alias(const char *key) {
 }
 
 bool set_timeout(const char *value) {
-    // TODO
-    (void)value;
+    int new_value = 0;
+    if (!parse_string_to_int(value, &new_value)) return false;
+    shell_config.max_history_len = new_value;
     return true;
+}
+
+bool unset_timeout() {
+    shell_config.timeout = -1;
+    return true;
+}
+const char *get_timeout() {
+    return int_to_string(shell_config.timeout);
 }
 
 bool set_show_cmd_path(const char *value) {
@@ -177,7 +216,106 @@ const char *get_env(const char *key) {
     return get__env(key);
 }
 
+bool set_terminal_mode(const char *value) {
+    TerminalMode mode = get_mode(value);
+    if (mode == TERM_MODE_UNKOWN) {
+        shell_print(SHELL_WARN, "Unknown Terminal Mode Value: %s", value);
+        shell_print(SHELL_WARN, "Valid Modes Are:\r\n    `normal`\r\n    `vi`\r\n    `emacs`", value);
+        return false;
+    }
+    shell_config.terminal_mode = mode;
+    return true;
+}
+
+bool unset_terminal_mode() {
+    shell_config.terminal_mode = TERM_MODE_NORMAL;
+    return true;
+}
+
+const char *get_terminal_mode() {
+    return mode_name(shell_config.terminal_mode);
+}
+
+bool set_enable_history_file(const char *value) {
+    shell_config.enable_history_file = parse_bool(value, shell_config.enable_history_file);
+    return true;
+}
+
+bool unset_enable_history_file() {
+    shell_config.enable_history_file = true;
+    return true;
+}
+
+const char *get_enable_history_file() {
+    return shell_config.enable_history_file ? "true" : "false";
+}
+
+bool set_enable_history(const char *value) {
+    shell_config.enable_history= parse_bool(value, shell_config.enable_history);
+    return true;
+}
+
+bool unset_enable_history() {
+    shell_config.enable_history= true;
+    return true;
+}
+
+const char *get_enable_history() {
+    return shell_config.enable_history ? "true" : "false";
+}
+
+bool set_max_history_len(const char *value) {
+    int new_value = 0;
+    if (!parse_string_to_int(value, &new_value)) {
+        shell_print(SHELL_WARN, "Failed to parse %s to an int\n", value);
+        return false;
+    }
+    shell_config.max_history_len = new_value;
+    return true;
+}
+
+bool unset_max_history_len() {
+    shell_config.max_history_len = 4096;
+    return true;
+}
+
+const char *get_max_history_len() {
+    return int_to_string(shell_config.max_history_len);
+}
+
+bool set_history_file(const char *value) {
+    char path[PATH_MAX] = {0};
+
+    if (!expand_path(value, path, PATH_MAX)) {
+        shell_print(SHELL_WARN, "Failed to expand path, please provide a valid one");
+        return false;
+    }
+
+
+    char *new_path = strdup(path);
+    if (!new_path) {
+        shell_print(SHELL_WARN, "Failed to allocate memory for history file path");
+        return false;
+    }
+
+    free(shell_config.history_file);
+    shell_config.history_file = new_path;
+    return true;
+}
+bool unset_history_file() {
+    shell_config.history_file = "~/.shellhistory";
+    return true;
+}
+const char *get_history_file() {
+    return shell_config.history_file;
+}
+
 bool init_shell_config() {
+    shell_config.history_file = strdup("~/.shellhistory");
+    if (!shell_config.history_file) {
+        shell_print(SHELL_ERROR, "FAiled to set history file path\n");
+        exit(1);
+    }
     size_t i;
     for (i = 0; i < MAX_ENVS; i++) {
         if (shell_config.envs[i].key[0] == '\0') {