+++ src/core/repl.h
@@ -1,6 +1,8 @@
 #ifndef STEIN_SHELL_REPL_H
 #define STEIN_SHELL_REPL_H
 
+#define MAX_LINE_LEN 1024
+
 void repl_loop();
 
 #endif // STEIN_SHELL_REPL_H